src/train/requirements.txt
boto3>=1.40.0
duckdb>=1.4.0
joblib>=1.4.0
numpy>=2.0.0
pandas>=2.0.0
pyarrow>=15.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
scikit-learn>=1.5.0,<2.0.0
skl2onnx>=1.17.0
onnxmltools>=1.12.0
onnxruntime>=1.20.0
xgboost>=2.1.0,<3.0.0
